Refactor app context to use redux, AuthGuard and routing
Replace the existing application context with Redux to manage the application state. Add protected routes to the application to prevent logged in users from accessing the login and register pages. When a user tries to access these routes while logged in, they will be redirected to the home page instead.