Move supabase file, fetch data from database and change styles in section
The Supabase file is a library used by the application for interfacing with the supabase database. In Home page fetched playlists from database to display them in Section comoponent. Fix div with song class. The icon of album was incorrect displayed when the title of the song was too long thats why I change display flex to grid and set width for title